requirements.txt
Flask~=1.1.2
pyenchant~=3.1.1
langdetect~=1.0.8